Flow-Solver-Matlab
This a single script in MATLAB to solve Navier-Stokes Equation with Convection, diffusion and pressure correction term. 
The grid is 100x100 2D and the velocity fields used are from Taylor-Green-Vortex in 2D. 
The code can be changed from Navier Stokes to any other form of transport equation by changing the value of phi and removing the pressure correction term accordingly.